Split get_query_result_set() into two in utils.py
Separated the get_query_result_set() method into parse_request to handle
parsing only, and another method to retrieve the query_result_set. This
separates the concerns in the old method.

Also added new functionality to the utils.py where we will now classify
the query_result_set into two baskets, with GPS and without GPS and
return it to the client.